+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 UNRELEASED
 ------------------
 - switch from `blue-tape` to `tape` and `tape-promise`
+- rename `global()` to `globalMarket()`
 
 
 0.1.0 / 2017-04-12